Code review & refactor 1
