+++ server.js
@@ -4,6 +4,11 @@ const bcrypt = require('bcrypt-nodejs');
 const cors = require('cors');
 const knex = require('knex');
 
+const register = require('./controllers/register');
+const signin = require('./controllers/signin');
+const profile = require('./controllers/profile');
+const image = require('./controllers/image');
+
 const db = knex({
   client: 'pg',
   connection: {
@@ -23,82 +28,13 @@ app.get('/', (req, res) => {
   res.send(database.users);
 });
 
-app.post('/signin', (req, res) => {
-  const { email, password } = req.body;
+app.post('/signin', (req, res) => { signin.handleSignIn(req, res, db, bcrypt) });
 
-  db.select('email', 'hash').from('login')
-  .where( 'email', '=', email)
-    .then(signInData => {
-      const isValid = bcrypt.compareSync(password, signInData[0].hash)
-      if (isValid) {
-        return db.select('*').from('users')
-          .where('email', '=', email)
-          .then(user => {
-            res.json(user[0]);
-          })
-          .catch(error => res.status(400).json('Unable to get user'));
-      } else {
-        res.status(400).json('Invalid credentials');
-      }
-    })
-    .catch(error => res.status(400).json('Invalid credentials'));
-});
+app.post('/register', (req, res) => { register.handleRegister(req, res, db, bcrypt) });
 
-app.post('/register', (req, res) => {
-  const { email, name, password } = req.body;
-  const hash = bcrypt.hashSync(password);
+app.get('/profile/:id', (req, res) => { profile.handleProfileGet(req, res, db) });
 
-  db.transaction(trx => {
-    trx.insert({ hash, email })
-      .into('login')
-      .returning('email')
-      .then(loginEmail => {
-        return trx('users')
-          .returning('*')
-          .insert({
-            email: loginEmail[0],
-            name,
-            joined: new Date()
-          })
-          .then(user => {
-            res.json(user[0])
-          })
-      })
-      .then(trx.commit)
-      .catch(trx.rollback);
-  })
-  .catch(error => res.status(400).json('Unable to register'));
-});
-
-app.get('/profile/:id', (req, res) => {
-  const { id } = req.params;
-
-  db.select('*').from('users').where({ id })
-    .then(user => {
-      if (user.length) {
-        res.json(user[0]);        
-      } else {
-        res.status(404).json('User not found')
-      }
-    })
-    .catch(error => { res.status(400).json('Error getting user') });
-});
-
-app.put('/image', (req, res) => {
-  const { id } = req.body;
-  
-  db('users').where('id', '=', id)
-  .increment('entries', 1)
-  .returning('entries')
-  .then(entries => {
-    if (entries.length) {
-        res.json(entries[0]);        
-    } else {
-      res.status(404).json('User not found')
-    }
-  })
-  .catch(error => { res.status(400).json('Error getting entries') });
-});
+app.put('/image', (req, res) => { image.handleImage(req, res, db) });
 
 app.listen(3000, () => {
   console.log('App is running on port 3000');
